--- pubnub-entities/src/main/kotlin/com/pubnub/entities/model/PNRemoveSpaceResult.kt
@@ -1,5 +1,0 @@
-package com.pubnub.entities.model
-
-class PNRemoveSpaceResult(
-    val status: Int
-)
