--- pubnub-entities/src/main/kotlin/com/pubnub/entities/model/PNRemoveUserResult.kt
@@ -1,5 +1,0 @@
-package com.pubnub.entities.model
-
-data class PNRemoveUserResult(
-    val status: Int
-)
