--- pubnub-entities/src/main/kotlin/com/pubnub/entities/model/PNSpace.kt
@@ -1,10 +1,0 @@
-package com.pubnub.entities.model
-
-data class PNSpace(
-    val id: String,
-    val name: String?,
-    val description: String?,
-    val custom: Any?,
-    val updated: String?,
-    val eTag: String?
-)
