--- pubnub-entities/src/main/kotlin/com/pubnub/entities/model/PNSpaceResult.kt
@@ -1,7 +1,0 @@
-package com.pubnub.entities.model
-
-data class PNSpaceResult(
-    val status: Int,
-    val data: PNSpace?
-)
-
