--- pubnub-entities/src/main/kotlin/com/pubnub/entities/model/PNUserArrayResult.kt
@@ -1,11 +1,0 @@
-package com.pubnub.entities.model
-
-import com.pubnub.api.models.consumer.objects.PNPage
-
-data class PNUserArrayResult(
-    val status: Int,
-    val data: Collection<PNUser>,
-    val totalCount: Int?,
-    val next: PNPage.PNNext?,
-    val prev: PNPage.PNPrev?
-)
