--- pubnub-entities/src/main/kotlin/com/pubnub/entities/model/PNUserResult.kt
@@ -1,6 +1,0 @@
-package com.pubnub.entities.model
-
-data class PNUserResult(
-    val status: Int,
-    val data: PNUser?
-)
