+++ pubnub-entities/src/main/kotlin/com/pubnub/entities/models/consumer/space/PNSpaceResult.kt
@@ -1,0 +1,16 @@
+package com.pubnub.entities.models.consumer.space
+
+import com.pubnub.api.models.consumer.objects.channel.PNChannelMetadataResult
+
+data class PNSpaceResult(
+    val status: Int,
+    val data: PNSpace?
+)
+
+internal fun PNChannelMetadataResult?.toPNSpaceResult(): PNSpaceResult? {
+    if (this == null) {
+        return null
+    }
+
+    return PNSpaceResult(status = status, data = data?.toPNSpace())
+}
