+++ src/main/kotlin/com/pubnub/api/models/consumer/objects/channel/PNChannelMetadataResult.kt
@@ -1,0 +1,6 @@
+package com.pubnub.api.models.consumer.objects.channel
+
+data class PNChannelMetadataResult(
+    val status: Int,
+    val data: PNChannelMetadata?
+)
