+++ src/main/kotlin/com/pubnub/api/models/consumer/objects/uuid/PNUUIDMetadataResult.kt
@@ -1,0 +1,6 @@
+package com.pubnub.api.models.consumer.objects.uuid
+
+data class PNUUIDMetadataResult(
+    val status: Int,
+    val data: PNUUIDMetadata?
+)
